Receiving Generic Errors from TransQgwDbEngine Server
Quantum Gateway does error handling in a non-standard way. Whenever ANY transaction with the server goes through, whether it's a success or failure, Quantum Gateway's TransQgwDbEngine Server sends an OK 200 code.

This means we have to handle the errors that they send differently.

Rather than catching 400 style errors as an HTTP response, to find errors from quantum gateway, we accept all responses from QGW since they're 200, but then we have to parse the returned response string to figure out if it's an error or not.

Generic Errors from Quantum Gateway like:
sending a request with a non existent or incorrect gateway login (basically Access API Key), sends a generic error message as a string.

What sets this message apart from other types of success or error responses is that it doesn't have APPROVED or DECLINED in the string.

So whenever we get a message from QGW without those two words, we know that QuantumGateway server is telling us that something went wrong before the transaction could even get processed. This is a generic error.