TransQGW Server Declined Transaction Error
Engine catches instances of TransactionError

One possible TransactionError it could catch has code:
ERR_SERVER_DECLINED

This means that a TransQGWDb Engine transaction got processed on the server, but there was some issue with the formatting of the request.

So our request gets a response back with a string like this:
"DECLINED","019452","65735","Y","M","0.3","Auth Declined","200

When we post our transaction to the TransQGWDbEngine in the postToServer function, we catch and throw this type of declined error when we call postToServer

in postToServer, we call the catchQuantumGatewayErrors function, which parses the declined string with parsePipeDelimitedResponse function, and then we create a TransctionResponse object with code: ERR_SERVER_DECLINED.
This object includes in its issues field a TransactionDeclinedIssue which is a CustomIssue. The TransactionDeclinedIssue holds a field called serverResponse which is a separate TransactionResponse object detailing what errors happened in the Server's response.

Finally,
the TransparentDbEngine catches and throws the TransactionError with all of the info collected from the postToServer error handling step.