TSDoc Param Warning Fix:
Error:
[warning] The signature src.TransactionError.constructor has an @param with name "message", which was not used
[warning] The signature src.TransactionError.constructor has an @param with name "issues", which was not used
[warning] The signature src.TransactionError.constructor has an @param with name "code", which was not used

the issue is with how TypeDoc is interpreting the destructured parameters in the constructor. When you use object destructuring in the parameters, TypeDoc expects the parameter documentation to match the structure of the destructured object.

to put simply, TSDoc interprets the incorrect doc comment:
@param message,
@param issues,
@param code
as three separate items in the parameter.

When in fact, what we have is a single item in the parameter, a destructured object (notice the curly braces):
{
message,
issues,
code,
}

To fix this, we make a single reference to the destructured object:
@param options
once we've captured that destructured object as a param, we can then reference its members:
options.message
options.issues
etc.